Reorganized pytest's fixtures
