+++ joboffers/tests/test_views.py
@@ -4,9 +4,10 @@ from ..models import JobOffer, OfferState
 from .factories import JobOfferCommentFactory, JobOfferFactory
 # Fixtures
 from .fixtures import ( # noqa
-    create_client, create_user_company_profile, create_logged_client,
-    create_publisher_client, create_user
+    create_publisher_client
 )
+from pyarweb.tests.fixtures import create_client, create_logged_client, create_user # noqa
+from pycompanies.tests.fixtures import create_user_company_profile # noqa
 #
 from django.contrib.messages import get_messages as contrib_get_messages
 from django.urls import reverse
